Move some stuff from BrowserMain into helper functions.
This patch changes the behavior slightly. We used to skip creation of the
metrics service at all if it was disabled on the command line. But this was
pretty confusing because it was still possible for it to be disabled by way of
not uploading any data. And there were two places in the rest of the function
that would crash if the metrics service was NULL (because you used the command
line switch).

After some archaeology, I discovered the command line switch was from before
we had preferences and was then the only way to disable the metrics. Since we
have prefs now and it defaults to off, I think the command line switch isn't
useful, and the fact that it crashes sometimes further indicates this. So I
just removed it.

This does not remove the actual constant since that's used by ChromeFrame:
once to make sure it's passed through when Chrome is spawend, and once in a
test that command line switches are passed through. We'll have to remove these
in a separate patch before removing the constant declaration.

This also moves the WebSocketExperimentRunner starting from outside the
"disable metrics" block. I think the intent was that it wouldn't be enabled
if we weren't collecting metrics, but it wasn't actually working that way,
so I just pulled it out to the top level of BrowserMain.

TEST=none
BUG=none
Review URL: http://codereview.chromium.org/1251007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@42629 0039d316-1c4b-4281-b951-d872f2087c98