Add the new extension install dialog for mac.
There are two xibs in this change.

ExtensionInstallPrompt.xib: This handles the more common case of installing an extension with one or more permission warnings. There is an NSBox with room for the warnings which are separated by \n\n.

ExtensionInstallPromptNoWarnings.xib: It is also possible to install an extension with zero permission warnings. This is currently rare, but we hope to make it more common in the coming quarters. In this case, there is no NSBox and the layout is slightly different.

BUG=29633

Review URL: http://codereview.chromium.org/1540013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@43575 0039d316-1c4b-4281-b951-d872f2087c98