Theme delegate fixes for the 10.6 SDK folllowing r43708.
In the 10.6 SDK, -[NSWindow delegate] returns id<NSWindowDelegate>, not just
id, and the three selectors used in r43708 are not declared in
NSWindowDelegate.  This results in errors such as "warning: '-themeProvider'
not found in protocol(s)".

Testing the selectors before using them is safe and not incorrect, but you
guys might actually want to make more of an assertion about what it means to
be a ChromeBrowserWindow's or FullscreenWindow's delegate, or perhaps even
a ChromeEventProcessingWindow's delegate.  Alternatively, it may be
appropriate to add a ChromeThemedWindow layer as a subclass of CEPW and
superclass of CBW and FW.  (CEPW's other subclass is InfoBubbleWindow.)

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/1630001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@43746 0039d316-1c4b-4281-b951-d872f2087c98