Fix flickering mac extension bubbles
The extension bubble animation would trigger an infinite loop of resizing of the web content between two heights.  By deferring the call to render_widget_host->WasResized() it seems to allow the height to settle and gets rid of the flickering.  This was mostly prevalent on 10.5, but occasionally showed up on 10.6 when using devtools to inspect the popup.

BUG=31970
TEST=Using RSS extension, open the rss popup at www.aaronboodman.com

Review URL: http://codereview.chromium.org/1596013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@44413 0039d316-1c4b-4281-b951-d872f2087c98