Free transport DIBs and associated shared memory when a WebPluginDelegateProxy is destroyed.
On the Mac, transport DIBs are allocated by the browser process (due to the
sandbox) and are also freed by the browser process. This change causes the
renderer to ask the browser to free allocated transport DIBs when a
WebPluginDelegateProxy is destroyed. This avoids a shared memory leak and
file descriptor leak in the browser process.

BUG=38383
TEST=Using test_chrome_fd_leak from bug 38383 and lsof, the browser process
     should not show an fd leak on Mac.
Review URL: http://codereview.chromium.org/1789006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@45940 0039d316-1c4b-4281-b951-d872f2087c98