linux: move plugin blacklist check earlier
Plugins load in two stages: first we dlopen() them to query metadata,
then later in response to a page request we actually instantiate them.
The blacklist we had in place only happened at the latter point;
gecko-mediaplayer causes us to hang in the first step.  So move the
blacklist check earlier to succesfully blacklist it.

We eventually should just fix the bugs that cause this plugin to hang us,
but this will fix things for users right now.

BUG=24507,40221
TEST=run with --log-level=0 --debug-plugin-loading and verify gecko-mediaplayer is still blacklisted

Review URL: http://codereview.chromium.org/2114011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@47656 0039d316-1c4b-4281-b951-d872f2087c98