Restricting lifetime of python sync server on Windows via a JobObject.
If a sync integration test runs for 30 seconds or more, OutOfProcTestRunner
forcibly kills the test case executable. This leaves an orphaned python
server instance in memory. When a subsequent test case is run, it
attempts to kick off a new python server and immediately checks for the
existence of a running server, following which it pumps messages to the
server instance. If the server detected by the test case happens to be
the old orphaned instance (that goes on to die when a new server
instance is started), test cases can fail.

This change list restricts the lifetime of the python.exe process started by
a test case such that if the test case is killed by OutOfProcTestRunner,
the python server instance it created dies with it.

In order to restrict the lifetime of a test server, it needs to be
started as a job. When a test server is spawned as a child process by a
test case that is running under a debugger, the child process needs to
be created using the CREATE_BREAKAWAY_FROM_JOB flag to first
disassociate it from the JobObject created by the debugger.

TestServerLauncher::Start() used to invoke base::LaunchApp() in order to
create the child process. This changelist implements a new method in
class TestServerLauncher called LaunchTestServerAsJob(). The call to
base::LaunchApp() in TestServerLauncher::Start() is replaced with a call
to TestServerLauncher::LaunchTestServerAsJob().

BUG=43777
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/2344001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@48762 0039d316-1c4b-4281-b951-d872f2087c98