Test end-to-end startup performance of loading several tabs.
This test is similar to other startup tests, with the exception that it loads multible tabs at once, and it tests both with and without the session restore command line flag. In both cases, WaitForInitialLoads waits for all the tabs to finish loading before returning.

BUG=44129
TEST=none

Review URL: http://codereview.chromium.org/2753001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@49827 0039d316-1c4b-4281-b951-d872f2087c98