Fix a potential crash in MountLibraryImpl::Init().
This is not a big deal, as the crash does not occur on the netbook.
I encountred a crash when I was trying to run chrome for chrome os with
libcros.so on my Ubuntu workstation. Note that building libcros.so outside
of chroot is very tricky, so no one else probably wants to do that.

TEST=manually
BUG=none

Review URL: http://codereview.chromium.org/2884015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@51915 0039d316-1c4b-4281-b951-d872f2087c98