Work around exception in command "which" while running GYP inside ChromiumOS chroot.
ChromiumOS is trying to remove dpkg-buildpackage.
If running inside ChromiumOS chroot, the "which" can throw exception when the target file is missing.
( dpkg-buildpackage will be missing in this case )
This doesn't happen with the "which" on Ubuntu.

Work around it by using "whereis" instead. It works on both ChromiumOS choot and Ubuntu.

Review URL: http://codereview.chromium.org/2945011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@52099 0039d316-1c4b-4281-b951-d872f2087c98