Gaia authentication for access remoting directory and talk
We are now able to access remoting directory token and talk token through
the chromoting setup dialog.

These information are sent to the service process for enabling the chromoting
host and perform the host registration.

All the plumbing for starting the service process, enabling the chromoting
host through through IPC commands are in place but host registration is
stubbed out and will be completed in next patch.

Review URL: http://codereview.chromium.org/3176014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@56278 0039d316-1c4b-4281-b951-d872f2087c98